authentication/app/logic/create_tokens.py
from datetime import datetime, timedelta
from jose import JWTError, jwt

SECRET_KEY = "f9e33dc7285d389df12b0dd47463083abcfe29a0c20a8f86c8fafebe7505d177"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
